requirements.txt
fastapi[all]
uvicorn
python-dotenv
firebase-admin
vertexai
google-cloud-aiplatform
google-cloud-storage
mysql-connector-python
pymysql
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
alembic
pytest
websockets==14.2
asyncio==3.4.3
google-auth

